Mitigate negative in-transit
For an unknown reason, simutrans sometimes records more goods being
delivered than were expected to be delivered.  This leads to a negative
in-transit amount.  Since the game calculates whether more goods are
allowed to be shipped from a supplier by seeing if the currently
in-transit amount is less than the maximum in transit amount, a negative
in-transit inflates the amount of goods that can be sent.

Note that this doesn't solve the reason that the game is recording more
goods delivered than expected, it merely mitigates the problem once it
occurs.